created a stylesheet for the product-list component and added it into the component.
created a custom interface for the product data to ensure all data conforms.

created a custom pipe to convert any provided character into a 'space character'.

updated the product-list component to allow for actual text filtering on the data-set by creating a getter and setter and a custom function to perform the filter.  This also needed an update to the view for product-list to use the new filter variable.

 added in OnInit for later use.