Setting up the ability to call data from a server rather than hard code: Added in the HttpClientModule to the app.module
added in the HttpClient, Observable and catch and do functions to the product service. Setup the constructor and revised the getProducts function to pull in data from a JSON file for now. Also setup error handling basics.

added in the subscribe to the ngOnInit to retrieve the data and setup basic error handling.